Start on fixing ERRORLEVEL check
So far I've just added a case that reliably does the wrong thing
(to be removed later) and fixed the related but separate bug that
the variables in :push_in were not local.

Without the latter change (adding setlocal), the value from the
previous call is used when the "for /f" loop has zero iterations.
This way, an empty value is used. (If the calling environment is
itself cmd.exe and a "value" variable is set, that might not be
the case.)